Revert "Use session storage instead of cookies for site notices"
This reverts commit 4ffa911964190041110bbc9e3b8ac10bd235d7bf.

Unlike what we thought, sessionStorage is not limited to the browsing
session (like session cookies are), but rather, it is limited to the
browsing context of a single tab/window. It does persist on reloads
and upon following links within the same tab, but not otherwise.

That makes is unsuitable for this data. Until we find a better solution,
let's stick to cookies for now.

Bug: T199274
Change-Id: Iaf0ac4e5458a7405b11601a2b9fbf594eeaf12aa